+++ src/environments/environment.failtests.ts
@@ -6,5 +6,5 @@
 export const environment = {
   production: false,
   googletranslate_api_key: null,
-  googletranslate_provoke_failures: false
+  googletranslate_provoke_failures: true
 };